# integrations/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ChangeX API",
    "description": "Provenance-first tracked editing for Word documents, over HTTP. Open a .docx for tracked editing, discover node ids via the outline, make the SMALLEST possible intent-named edit per call (always passing the exact existing text in `before` — blind overwrites are refused), then save a native Word accept/reject revisions file plus a hash-chained .changex provenance journal. A passive (no-tool-calling) /open + /seal path serves offline models. This schema is consumable directly as a ChatGPT custom GPT Action.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/sessions": {
      "post": {
        "tags": [
          "tracked"
        ],
        "summary": "Open Session",
        "description": "Open a .docx for tracked editing and return a session handle + summary.",
        "operationId": "openTracked",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenTrackedRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenTrackedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{handle}/outline": {
      "get": {
        "tags": [
          "tracked"
        ],
        "summary": "Get Outline",
        "description": "Return a bounded, paginated outline of the document's paragraphs.",
        "operationId": "getOutline",
        "parameters": [
          {
            "name": "handle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Handle"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque pagination cursor.",
              "title": "Cursor"
            },
            "description": "Opaque pagination cursor."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Max entries per page.",
              "default": 100,
              "title": "Limit"
            },
            "description": "Max entries per page."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutlineResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{handle}/edit": {
      "post": {
        "tags": [
          "tracked"
        ],
        "summary": "Edit",
        "description": "Apply ONE small, intent-named tracked edit to a node and journal it.",
        "operationId": "editSession",
        "parameters": [
          {
            "name": "handle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Handle"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{handle}/save": {
      "post": {
        "tags": [
          "tracked"
        ],
        "summary": "Save",
        "description": "Save the native-revisions .docx and report the tracked + .changex paths.",
        "operationId": "saveSession",
        "parameters": [
          {
            "name": "handle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Handle"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{handle}/changes": {
      "get": {
        "tags": [
          "tracked"
        ],
        "summary": "Get Changes",
        "description": "Return the structured provenance journal (active, non-reverted events).",
        "operationId": "getChanges",
        "parameters": [
          {
            "name": "handle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Handle"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/open": {
      "post": {
        "tags": [
          "passive"
        ],
        "summary": "Passive Open",
        "description": "Snapshot a docx and write a pending passive journal (no tool calls).\n\nAny tool may then edit the docx freely; call /seal to capture the delta.",
        "operationId": "passiveOpen",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PassiveOpenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PassiveOpenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/seal": {
      "post": {
        "tags": [
          "passive"
        ],
        "summary": "Passive Seal",
        "description": "Diff the edited docx vs the stored baseline and append passive ops.\n\nReturns honest, *degraded* counts: passive ops are observed net textual\ndeltas, not true provenance (agent/vendor/prompt are null).",
        "operationId": "passiveSeal",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PassiveSealRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PassiveSealResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/report": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Report",
        "description": "Render an HTML/markdown redline of a session's or journal's changes.\n\nSupply either an open ``handle`` (tracked flow) or a ``changex`` journal\npath (passive flow). With ``raw=true`` and ``fmt=html`` the response is a\nready-to-display ``text/html`` page; otherwise a JSON ``{format, report}``.",
        "operationId": "renderReport",
        "parameters": [
          {
            "name": "handle",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "An open tracked-session handle to render from.",
              "title": "Handle"
            },
            "description": "An open tracked-session handle to render from."
          },
          {
            "name": "changex",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "OR a path to a .changex journal to render from (passive flow).",
              "title": "Changex"
            },
            "description": "OR a path to a .changex journal to render from (passive flow)."
          },
          {
            "name": "fmt",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "'html' (default) or 'markdown'.",
              "default": "html",
              "title": "Fmt"
            },
            "description": "'html' (default) or 'markdown'."
          },
          {
            "name": "raw",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If true, return the report as text/html instead of JSON.",
              "default": false,
              "title": "Raw"
            },
            "description": "If true, return the report as text/html instead of JSON."
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Renderreport"
                }
              },
              "text/html": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OpenTrackedRequest": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path",
            "description": "Absolute server-side path to the .docx to open for tracked editing.",
            "examples": [
              "/data/contract.docx"
            ]
          },
          "agent_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentContext"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional once-per-session declared identity {model, vendor}."
          },
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author",
            "description": "Override the Word revision author (defaults to the model id)."
          }
        },
        "type": "object",
        "required": [
          "path"
        ],
        "title": "OpenTrackedRequest",
        "description": "Body for POST /sessions — open a .docx for tracked editing."
      },
      "OpenTrackedResponse": {
        "properties": {
          "handle": {
            "type": "string",
            "title": "Handle",
            "description": "Opaque session handle; pass it to every other call."
          },
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "description": "Stable journal session id (provenance identity)."
          },
          "baseline_sha256": {
            "type": "string",
            "title": "Baseline Sha256",
            "description": "SHA-256 of the opened baseline bytes."
          },
          "summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Summary",
            "description": "{filename, paragraphs, agent, vendor, changex_path}."
          }
        },
        "type": "object",
        "required": [
          "handle",
          "session_id",
          "baseline_sha256",
          "summary"
        ],
        "title": "OpenTrackedResponse",
        "description": "Session handle + summary returned by POST /sessions."
      },
...
      "OutlineResponse": {
        "properties": {
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/OutlineEntry"
            },
            "type": "array",
            "title": "Nodes"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor",
            "description": "Pass back as ?cursor= to page forward; null when done."
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total paragraph count in the document."
          }
        },
        "type": "object",
        "required": [
          "nodes",
          "total"
        ],
        "title": "OutlineResponse",
        "description": "A bounded outline page plus the cursor to fetch the next one."
      },
...
    }
  }
}